refactor: work in progress
