+++ src/scraper/characterData/characterSkills.js
@@ -1,5 +1,5 @@
-import { rawSkillData } from './rawSkillData.js';
-import fs from 'fs';
+// import { rawSkillData } from './rawSkillData.js';
+// import fs from 'fs';
 
 /* helper functions to format data */
 
@@ -61,11 +61,11 @@ function transformCharacterData(data) {
   });
 }
 
-let data = await transformCharacterData(rawSkillData);
-fs.writeFile('output.json', JSON.stringify(data), (err) => {
- if (err) throw err;
-});
+// let data = await transformCharacterData(rawSkillData);
+// fs.writeFile('output.json', JSON.stringify(data), (err) => {
+//  if (err) throw err;
+// });
 
-console.log(transformCharacterData(rawSkillData))
+// console.log(transformCharacterData(rawSkillData))
 
 export { transformCharacterData };
