+++ src/scraper/commandListPage/processCommandListPage.js
@@ -1,0 +1,20 @@
+import { getCommandPageData } from './commandScrape.js'
+
+
+async function processCommandListPage(url) {
+  
+  console.log(`Processing ${url}`);
+
+  const browser = await puppeteer.launch({ headless: true, args: ['--no-sandbox'] });
+  const page = await browser.newPage();
+  await page.setUserAgent('Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.190 Safari/537.36');
+  await page.goto(url);
+
+  const data = await page.evaluate(`(${getCommandPageData.toString()})();`);
+
+  await browser.close();
+  return data;
+}
+
+
+export { processCommandListPage }
