+++ src/scraper/frameDataPage/processFrameDataPage.js
@@ -1,0 +1,28 @@
+import { getVitality } from './vitalityValue.js'
+import { extractDataFromTable } from './tableScrape.js';
+
+async function processFrameDataPage(url) {
+  
+  console.log(`Processing ${url}`);
+
+  const browser = await puppeteer.launch({ headless: true, args: ['--no-sandbox'] });
+
+  const page = await browser.newPage();
+  await page.setUserAgent('Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.190 Safari/537.36');
+  
+  await Promise.all([
+    page.goto(url),
+    page.waitForNavigation({ waitUntil: 'networkidle0' }),
+  ]);
+
+  const data = await page.evaluate(`(${extractDataFromTable.toString()})()`);
+  const vitality = await page.evaluate(`(${getVitality.toString()})()`);
+
+  await browser.close();
+  return {
+    data,
+    vitality
+  };
+};
+
+export { processFrameDataPage };
